requirements.txt
asgiref==3.5.2
autopep8==1.7.0
awscli==1.25.83
black==22.12.0
botocore==1.27.82
click==8.1.3
colorama==0.4.4
dj-database-url==2.0.0
Django==4.1.4
django-ckeditor-5==0.2.4
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
docutils==0.16
gunicorn==20.1.0
jmespath==1.0.1
joblib==1.2.0
Markdown==3.4.1
mypy-extensions==0.4.3
numpy==1.24.1
pathspec==0.10.3
Pillow==9.4.0
platformdirs==2.6.2
pyasn1==0.4.8
pycodestyle==2.9.1
python-dateutil==2.8.2
pytz==2022.6
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.6.0
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.12
whitenoise==6.4.0
python-decouple
django-termsandconditions
validate-docbr